Modify: Change UI Layer Structure by Feature
