+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name=".ui.activity.MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             tools:ignore="LockedOrientationActivity">
@@ -31,24 +31,24 @@
         </activity>
 
         <activity
-            android:name=".ui.activity.SettingActivity"
+            android:name=".presentation.setting.SettingActivity"
             android:configChanges="orientation|screenSize|keyboardHidden"
             android:exported="true" />
 
         <activity
-            android:name=".ui.activity.ImageActivity"
+            android:name=".presentation.main.media.MediaActivity"
             android:configChanges="orientation|screenSize|keyboardHidden"
             android:exported="true" />
 
         <activity
-            android:name=".ui.activity.VideoActivity"
+            android:name=".presentation.main.media.video.VideoActivity"
             android:configChanges="orientation|screenSize|keyboardHidden"
             android:exported="true"
             android:theme="@style/AppTheme.Dark">
         </activity>
 
         <activity
-            android:name=".ui.activity.ExternalVideoActivity"
+            android:name=".presentation.main.media.video.ExternalVideoActivity"
             android:configChanges="orientation|screenSize|keyboardHidden"
             android:exported="true"
             android:theme="@style/AppTheme.Dark">
@@ -65,7 +65,7 @@
         </activity>
 
         <activity
-            android:name=".ui.activity.ChoiceActivity"
+            android:name=".presentation.choice.ChoiceActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             tools:ignore="LockedOrientationActivity" />
