+++ app/src/main/java/com/pleiades/pleione/slotgallery/controller/ContentController.kt
@@ -16,13 +16,10 @@ import com.pleiades.pleione.slotgallery.Config.Companion.PATH_SNAPSEED
 import com.pleiades.pleione.slotgallery.Config.Companion.SORT_POSITION_BY_NAME
 import com.pleiades.pleione.slotgallery.Config.Companion.SORT_POSITION_BY_NEWEST
 import com.pleiades.pleione.slotgallery.Config.Companion.SORT_POSITION_BY_OLDEST
-import com.pleiades.pleione.slotgallery.controller.ContentController.Companion.directoryArrayList
 import com.pleiades.pleione.slotgallery.info.Directory
 import com.pleiades.pleione.slotgallery.info.Slot
-import com.pleiades.pleione.slotgallery.ui.fragment.dialog.ProgressDialogFragment
-import kotlinx.coroutines.CoroutineScope
+import com.pleiades.pleione.slotgallery.presentation.dialog.ProgressDialogFragment
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import java.io.BufferedInputStream
 import java.io.BufferedOutputStream
