+++ app/src/main/java/com/pleiades/pleione/slotgallery/presentation/setting/ManageDirectoryFragment.kt
@@ -1,4 +1,4 @@
-package com.pleiades.pleione.slotgallery.ui.fragment.setting
+package com.pleiades.pleione.slotgallery.presentation.setting
 
 import android.content.Intent
 import android.media.MediaScannerConnection
@@ -22,7 +22,6 @@ import com.pleiades.pleione.slotgallery.Config.Companion.NAME_DUMMY
 import com.pleiades.pleione.slotgallery.Config.Companion.SETTING_POSITION_DIRECTORY
 import com.pleiades.pleione.slotgallery.R
 import com.pleiades.pleione.slotgallery.controller.SlotController
-import com.pleiades.pleione.slotgallery.databinding.FragmentMainBinding
 import com.pleiades.pleione.slotgallery.databinding.FragmentManageBinding
 import com.pleiades.pleione.slotgallery.info.Slot
 import java.io.BufferedInputStream
