+++ app/src/main/java/com/pleiades/pleione/slotgallery/presentation/setting/SettingFragment.kt
@@ -1,4 +1,4 @@
-package com.pleiades.pleione.slotgallery.ui.fragment.setting
+package com.pleiades.pleione.slotgallery.presentation.setting
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -15,7 +15,6 @@ import com.pleiades.pleione.slotgallery.Config.Companion.SETTING_POSITION_DIRECT
 import com.pleiades.pleione.slotgallery.Config.Companion.SETTING_POSITION_SLOT
 import com.pleiades.pleione.slotgallery.R
 import com.pleiades.pleione.slotgallery.controller.SlotController
-import com.pleiades.pleione.slotgallery.databinding.FragmentManageBinding
 import com.pleiades.pleione.slotgallery.databinding.FragmentSettingBinding
 
 class SettingFragment : Fragment() {
